refactor: split ContextsStats and ContextsInformers Signed-off-by: Philippe Martin <phmartin@redhat.com>
