+++ packages/main/src/plugin/kubernetes/contexts-states.spec.ts
@@ -15,86 +15,23 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
-import type { Informer, KubernetesObject } from '@kubernetes/client-node';
 import { describe, expect, test } from 'vitest';
 
 import { ContextsStates, isSecondaryResourceName } from './contexts-states.js';
-import { FakeInformer } from './kubernetes-context-state.spec.js';
 
 describe('ContextsStates tests', () => {
-  test('hasInformer should check if informer exists for context', () => {
-    const client = new ContextsStates();
-    client.setInformers(
-      'context1',
-      new Map([['pods', new FakeInformer('context1', '/path/to/resource', 0, undefined, [], [])]]),
-    );
-    expect(client.hasInformer('context1', 'pods')).toBeTruthy();
-    expect(client.hasInformer('context1', 'deployments')).toBeFalsy();
-    expect(client.hasInformer('context2', 'pods')).toBeFalsy();
-    expect(client.hasInformer('context2', 'deployments')).toBeFalsy();
-  });
-
-  test('getContextsNames should return the names of contexts as array', () => {
-    const client = new ContextsStates();
-    client.setInformers(
-      'context1',
-      new Map([['pods', new FakeInformer('context1', '/path/to/resource', 0, undefined, [], [])]]),
-    );
-    client.setInformers(
-      'context2',
-      new Map([['pods', new FakeInformer('context2', '/path/to/resource', 0, undefined, [], [])]]),
-    );
-    expect(Array.from(client.getContextsNames())).toEqual(['context1', 'context2']);
-  });
-
-  test('isReachable', () => {
-    const client = new ContextsStates();
-    client.setInformers(
-      'context1',
-      new Map([['pods', new FakeInformer('context1', '/path/to/resource', 0, undefined, [], [])]]),
-    );
-    client.setInformers(
-      'context2',
-      new Map([['pods', new FakeInformer('context2', '/path/to/resource', 0, undefined, [], [])]]),
-    );
-    client.safeSetState('context1', state => (state.reachable = true));
-
-    expect(client.isReachable('context1')).toBeTruthy();
-    expect(client.isReachable('context2')).toBeFalsy();
-    expect(client.isReachable('context3')).toBeFalsy();
-  });
-
   test('isSecondaryResourceName', () => {
     expect(isSecondaryResourceName('pods')).toBeFalsy();
     expect(isSecondaryResourceName('services')).toBeTruthy();
   });
 
-  test('informers registry', () => {
-    const states = new ContextsStates();
-    expect(states.hasContext('ctx1')).toBeFalsy();
-    expect(states.hasInformer('ctx1', 'services')).toBeFalsy();
-    expect(states.getContextsNames()).toMatchObject({});
+  test('isReachable', () => {
+    const client = new ContextsStates();
+    client.safeSetState('context1', state => (state.reachable = true));
 
-    states.setInformers('ctx1', new Map());
-    expect(states.hasContext('ctx1')).toBeTruthy();
-    expect(states.hasInformer('ctx1', 'services')).toBeFalsy();
-    expect(states.hasInformer('ctx1', 'pods')).toBeFalsy();
-
-    const informersWithService = new Map();
-    informersWithService.set('services', {} as Informer<KubernetesObject>);
-    states.setInformers('ctx1', informersWithService);
-    expect(states.hasContext('ctx1')).toBeTruthy();
-    expect(states.hasInformer('ctx1', 'services')).toBeTruthy();
-    expect(states.hasInformer('ctx1', 'pods')).toBeFalsy();
-
-    states.setResourceInformer('ctx1', 'pods', {} as Informer<KubernetesObject>);
-    expect(states.hasContext('ctx1')).toBeTruthy();
-    expect(states.hasInformer('ctx1', 'services')).toBeTruthy();
-    expect(states.hasInformer('ctx1', 'pods')).toBeTruthy();
-
-    expect(() => states.setResourceInformer('ctx2', 'pods', {} as Informer<KubernetesObject>)).toThrow(
-      'watchers for context ctx2 not found',
-    );
+    expect(client.isReachable('context1')).toBeTruthy();
+    expect(client.isReachable('context2')).toBeFalsy();
+    expect(client.isReachable('context3')).toBeFalsy();
   });
 
   test('state', () => {
